refactor: Improve settings migration
Key Changes:
- Moved `SETTINGS_SCHEMA_VERSION` and `DeprecatedSettingsSchema` to `shared.ts`. This resolves a circular dependency, ensuring `SETTINGS_SCHEMA_VERSION` is correctly defined when migration objects are initialized.
- Simplified conditional checks within individual migration functions.
- Modified `parseInfioSettings` to first attempt strict parsing of current settings, deferring to the migration process only when an older or non-standard format is detected.
